Improving performance radiators

+++ FNPlugin/GameConstants.cs
@@ -48,7 +48,7 @@ namespace FNPlugin
         public const double microwave_dish_efficiency = 0.85f;
         public const double microwave_alpha = 0.00399201596806387225548902195609f;
 
-        public const double stefan_const = 5.670373e-8;
+        public const float stefan_const = 5.670373e-8f;
         public const double warpspeed = 29979245.8;
         public const double rad_const_h = 1000;
         public const double alpha = 0.001998001998001998001998001998;
